+++ aws_efs/variables.tf
@@ -14,8 +14,12 @@ variable "subnet_id" {
 }
 
 variable "names" {
-  description = "Mapping of EFS names to their respective security groups"
-  type        = map(list(string))
+  type = list(object({
+    name             = string
+    security_groups  = list(string)
+    default_ap       = optional(bool, false)
+    throughput_mode  = optional(string, "bursting")
+  }))
 }
 
 variable "default_ap" {